agregando useEfect para las frases random
